@@ -42,17 +42,17 @@
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^9.0.0",
     "babel-jest": "^23.0.1",
-    "babel-loader": "^8.0.0-beta.0",
+    "babel-loader": "^8.0.5",
     "babel-plugin-module-resolver": "^3.1.1",
     "chalk": "^2.1.0",
-    "clean-webpack-plugin": "^1.0.0",
+    "clean-webpack-plugin": "^1.0.1",
     "css-loader": "^1.0.0",
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.2.0",
     "enzyme-to-json": "^3.3.5",
-    "eslint": "^5.3.0",
+    "eslint": "^5.12.1",
     "eslint-config-airbnb": "17.1.0",
-    "eslint-config-prettier": "^3.0.0",
+    "eslint-config-prettier": "^3.6.0",
     "eslint-import-resolver-babel-module": "5.0.0-beta.0",
     "eslint-plugin-flowtype": "^3.2.0",
     "eslint-plugin-import": "^2.14.0",
@@ -67,12 +67,12 @@
     "jest": "^23.1.0",
     "jest-styled-components": "^6.3.1",
     "jsdom": "^13.1.0",
-    "madge": "^3.0.1",
+    "madge": "^3.4.3",
     "minimist": "^1.2.0",
     "npm-run-all": "^4.1.3",
     "null-loader": "^0.1.1",
     "ora": "^3.0.0",
-    "prettier": "^1.15.3",
+    "prettier": "^1.16.1",
     "prettier-stylelint": "^0.4.2",
     "react-test-renderer": "^16.7.0",
     "style-loader": "^0.22.0",
